+++ server/app/models/user.js
@@ -87,18 +87,25 @@ UserSchema.path('username').validate(function(username) {
     if (authTypes.indexOf(this.provider) !== -1) return true;
     return username.length;
 }, 'Username cannot be blank');
+/*
 
 UserSchema.path('hashed_password').validate(function(hashed_password) {
     // if you are authenticating by any of the oauth strategies, don't validate
     if (authTypes.indexOf(this.provider) !== -1) return true;
     return hashed_password.length;
 }, 'Password cannot be blank');
+*/
 
 
 /**
  * Pre-save hook
  */
 UserSchema.pre('save', function(next) {
+    
+    this.fullname = this.name.first + ' ' + this.name.last; 
+    
+    console.log(this.fullname);
+    
     if (!this.isNew) return next();
 
     if (!validatePresenceOf(this.password) && authTypes.indexOf(this.provider) === -1)